templatize the opening crawl and play the music
This change inserts the data passed into the opening crawl into the
appropriate locations of the HTML template. It also introduces
localization for the into text.

Lastly, I'm introducing an audio tag to time the play of the music. I
used this in the iframe version, and so I'm bringing it over here for
completition. However, I suspect folks will complain if I don't
integrate with Foundry's audio system. If anyone has their playlist
volume down to 10%, they'll likely be annoyed by the intro music playing
at maximum volume.