merge opening crawl scene controls into one button
Wrycu wants a single button for each high level feature that uses the
scene controls. This commit merges the create new opening crawl journal
functionality into the opening crawl launcher. It renames the launcher
just Opening Crawl.

As part of this change, additional error checking and initialization
workflows are added. It will prompt to create the crawls folder. It will
also only show the launcher if the folder has entries within it.

I left the old journal button in place because we'll reuse it for the
datapad templates.

This closes #3.