remove buggy audio preload logic
This logic attempted to only play the music once the user had clicked on the window. This would allow audio to play correctly, since auto-play of music is disabled on a freshly loaded tab. My logic here was sketchy and attempted to hook something that clearly doesn't work the way I expected.