Change combatant name and icon in preCreate hook.
Updating embedded documents directly can lead to a race condition where
the update is sent to all connected clients, but not saved in the DB.
This leads to some icons and names reverting when reloading foundry. By
utilizing the preUpdate hook, the data of the Combatant can be edited
before sending the create request to the db, thus ensuring the changed
icon and name are stored in the db. The hook only runs on the client
that initiates the create request, eliminating the need to check that a
GM is running it, and preventing it from running multiple times if an
additional GM is logged in.

Additionally, this now handles Combatants created from an actor instead
of a token, making usage with the canvas disabled possible.