fix(tests): Fix race condition with closing initial popups in cypress tests
It turns out that FoundryVTT has a 500ms delay before their close
buttons become active to prevent accidental closures. Unfortunately, I
haven't yet found a way to wait for the event handler to be attached.

This change switches to clicking the accept buttons, which don't have
that delay.

This bug likely impacts other close buttons. (I have a double click to
work around this in an earlier step in initialization, but that's not
a blocker currently.)