tests(dice_helper): first pass at dice helper end-to-end test (#159)
Adds tests for the dice helper on both a Brawl and Ranged - Light attack. These two skills were chosen because we've sometime seen issues in generating helper results when there's a space in the skill name, so the mix of the two should help capture complete results.

As a dependency of this, the tests clear chat logs and create a test actor.