Fix usage of Attack Animation (Item)
The first entry of the list 'custom_item_animations' is not valid because it's the 'new entry' without information.
But it's always match with the item_name so the choosen animation is an empty one.
the two names are not lower case so the match is not exact.
i fix the log that use the list instead of the value found.

Thanks for your work it's awsome.